cleanup/reorganize: move projection logic out of 'queries'.
This allows the eventual split between XML parsing and ORM handling
