+++ gisserver/queries/__init__.py
@@ -13,7 +13,6 @@ The "GetFeatureById" is a mandatory built-in stored query.
 
 from .adhoc import AdhocQuery
 from .base import QueryExpression
-from .projection import FeatureProjection, FeatureRelation
 from .stored import (
     GetFeatureById,
     QueryExpressionText,
@@ -32,6 +31,4 @@ __all__ = (
     "stored_query_registry",
     "StoredQueryParameter",
     "GetFeatureById",
-    "FeatureProjection",
-    "FeatureRelation",
 )
